Refactor the loop of sliding window
The original loop requires 42(the outer loop) x 109(the inner loop) = 4578
iterations to search and sum the angles within the sliding pi/3 window.

Assuming the sorted list of angles, the new code only performs 42 iterations.
Even better, if the histogram of the angles is narrowly distributed, most
iterations will be skipped by continue.  The precision is slightly sacrificed
due to the rounding of quantum 0.15f.