Introduce compare_and_pack_descriptor()
This function writes the descriptor bits.  Instead of writing bits one by one,
it writes them by a "bucket" so that the number of writes is reduced to
1/(bucket size).

The byte sequence of the descriptor bits depends on the machine byte order
unless bucket is uint8_t.  In addition, the bit sequence of a byte is reversed,
compared to the original code.

See below for the correspondence of bits between the origianl and the new code
when bucket is uint8_t:

                   MSB           LSB         MSB                   LSB
  original desc[0] |7|6|5|4|3|2|1|0| desc[1] |15|14|13|12|11|10| 9| 8|
  new      desc[0] |0|1|2|3|4|5|6|7| desc[1] | 8| 9|10|11|12|13|14|15|