Change the first async task to be a deferred task.
By this change, the current thread rather than an async thread filters the
pixels of the first evolution level.  So, the current thread should have
enough work to finish before sleeping till the other async tasks end.

This change hopefully removes sleep and wakeup of the current thread due to
future::get() in this function.