Change Compute_Main_Orientation() a static non-member function
The function does not depends on the internal knowledge of AKAZEFeatures.
It is only referenced in AKAZEFeatures.cpp. Thus, a static non-member
function is more proper choice for this task.

In addition, the type of the argument "evolution_" is changed from
vector<TEvolution>& to TEvolution&, which is all this task needs.