Fix the regression due to the overuse of std::ref()
This bug is introduced by Support_CentOS7 and caused the issue #1 and #2.

Lx and Ly must be passed by value, instead of by reference, because Lx
and Ly are local variables and the lifetime of kcontrast_ async task
extends beyond the scope of Compute_Base_Evolution_Level().

Although the heap memory is supposed to be pinned down by
AKAZEFeatures::lx_ and ly_, its reference count seems to go wrong.