Fix a bug that Find_Scale_Space_Extrema() misses to filter out some points
The bug was injected by the following commit in R9.5.

  f0d437f7 Parallelize Find_Scale_Space_Extrema()

The code before f0d437f7 pushed a maximum point to kpts_aux vector
only when the point was stronger BOTH IN SPACE AND IN SCALE.

After f0d437f7, the code pushed all spatially maximum points
disregarding scale.  Non-extrema in scale must be deleted afterward.
However, the code now missed to delete a point when:

  - The point is weaker than the neighboring point of lower scale
  - The neighboring point is already deleted by the point of
    one more lower scale.

For example, if A,B,C are three neighboring points in three different
scales and their responses are A>B>C from lower to upper, filtering with
the upper scale level has deleted B due to A, but C will remain
because B is already deleted and not available even though A>C.

This change fixes the issue by filtering the points by two directions.
The direction from lower to upper filters out a point of LOWER SCALE ONLY
when its response is weaker than the current scale.  By the similar way,
the direction from upper to lower filters out a point of UPPER SCALE ONLY.

For the above example, the new code scans points from upper to lower
when comparing with the upper level.  So, C and B can be deleted safely
and A remains.

As a bonus the code has become more symmetric and easier.